chore: extract base controllers for all DSP apis
