+++ data-protocols/dsp/dsp-catalog/build.gradle.kts
@@ -23,4 +23,5 @@ dependencies {
     api(project(":data-protocols:dsp:dsp-catalog:dsp-catalog-transform"))
     api(project(":data-protocols:dsp:dsp-catalog:lib:dsp-catalog-validation-lib"))
     api(project(":data-protocols:dsp:dsp-catalog:lib:dsp-catalog-transform-lib"))
+    api(project(":data-protocols:dsp:dsp-catalog:lib:dsp-catalog-http-api-lib"))
 }
