+++ data-protocols/dsp/dsp-catalog/lib/dsp-catalog-http-api-lib/src/main/java/org/eclipse/edc/protocol/dsp/catalog/http/api/decorator/CatalogPaginationResponseDecorator.java
