+++ data-protocols/dsp/dsp-negotiation/build.gradle.kts
@@ -23,4 +23,5 @@ dependencies {
     api(project(":data-protocols:dsp:dsp-negotiation:dsp-negotiation-transform"))
     api(project(":data-protocols:dsp:dsp-negotiation:lib:dsp-negotiation-validation-lib"))
     api(project(":data-protocols:dsp:dsp-negotiation:lib:dsp-negotiation-transform-lib"))
+    api(project(":data-protocols:dsp:dsp-negotiation:lib:dsp-negotiation-http-api-lib"))
 }
