+++ data-protocols/dsp/dsp-negotiation/dsp-negotiation-http-api/build.gradle.kts
@@ -26,6 +26,7 @@ dependencies {
     api(project(":extensions:common:json-ld"))
 
     implementation(project(":data-protocols:dsp:dsp-negotiation:lib:dsp-negotiation-validation-lib"))
+    implementation(project(":data-protocols:dsp:dsp-negotiation:lib:dsp-negotiation-http-api-lib"))
     implementation(project(":extensions:common:http:lib:jersey-providers-lib"))
 
     implementation(libs.jakarta.rsApi)
@@ -33,6 +34,8 @@ dependencies {
     testImplementation(project(":core:common:junit"))
     testImplementation(testFixtures(project(":extensions:common:http:jersey-core")))
     testImplementation(libs.restAssured)
+    testImplementation(testFixtures(project(":data-protocols:dsp:dsp-negotiation:lib:dsp-negotiation-http-api-lib")))
+
 }
 
 edcBuild {
