+++ data-protocols/dsp/dsp-negotiation/lib/dsp-negotiation-http-api-lib/src/main/java/org/eclipse/edc/protocol/dsp/negotiation/http/api/NegotiationApiPaths.java
