+++ data-protocols/dsp/dsp-transfer-process/dsp-transfer-process-http-api/build.gradle.kts
@@ -26,12 +26,14 @@ dependencies {
 
     implementation(project(":spi:common:json-ld-spi"))
     implementation(project(":data-protocols:dsp:dsp-transfer-process:lib:dsp-transfer-process-validation-lib"))
+    implementation(project(":data-protocols:dsp:dsp-transfer-process:lib:dsp-transfer-process-http-api-lib"))
     implementation(project(":extensions:common:http:lib:jersey-providers-lib"))
 
     implementation(libs.jakarta.rsApi)
 
     testImplementation(project(":core:common:junit"))
     testImplementation(testFixtures(project(":extensions:common:http:jersey-core")))
+    testImplementation(testFixtures(project(":data-protocols:dsp:dsp-transfer-process:lib:dsp-transfer-process-http-api-lib")))
 
     testImplementation(libs.restAssured)
 }
