+++ settings.gradle.kts
@@ -88,6 +88,7 @@ include(":data-protocols:dsp:dsp-catalog:dsp-catalog-http-dispatcher")
 include(":data-protocols:dsp:dsp-catalog:dsp-catalog-transform")
 include(":data-protocols:dsp:dsp-catalog:lib:dsp-catalog-validation-lib")
 include(":data-protocols:dsp:dsp-catalog:lib:dsp-catalog-transform-lib")
+include(":data-protocols:dsp:dsp-catalog:lib:dsp-catalog-http-api-lib")
 include(":data-protocols:dsp:dsp-spi")
 include(":data-protocols:dsp:dsp-negotiation")
 include(":data-protocols:dsp:dsp-negotiation:dsp-negotiation-http-api")
@@ -95,6 +96,7 @@ include(":data-protocols:dsp:dsp-negotiation:dsp-negotiation-http-dispatcher")
 include(":data-protocols:dsp:dsp-negotiation:dsp-negotiation-transform")
 include(":data-protocols:dsp:dsp-negotiation:lib:dsp-negotiation-validation-lib")
 include(":data-protocols:dsp:dsp-negotiation:lib:dsp-negotiation-transform-lib")
+include(":data-protocols:dsp:dsp-negotiation:lib:dsp-negotiation-http-api-lib")
 include(":data-protocols:dsp:dsp-http-core")
 include(":data-protocols:dsp:dsp-http-spi")
 include(":data-protocols:dsp:dsp-transfer-process")
@@ -103,6 +105,7 @@ include(":data-protocols:dsp:dsp-transfer-process:dsp-transfer-process-http-disp
 include(":data-protocols:dsp:dsp-transfer-process:dsp-transfer-process-transform")
 include(":data-protocols:dsp:dsp-transfer-process:lib:dsp-transfer-process-validation-lib")
 include(":data-protocols:dsp:dsp-transfer-process:lib:dsp-transfer-process-transform-lib")
+include(":data-protocols:dsp:dsp-transfer-process:lib:dsp-transfer-process-http-api-lib")
 include(":data-protocols:dsp:dsp-version:dsp-version-http-api")
 include(":data-protocols:dsp:dsp-version:dsp-version-http-dispatcher")
